mcp_servers/python/servers/WHATSAPP/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsapp-mcp-server"
version = "0.2.0"
authors = [
    {name = "Vanij MCP Platform", email = "support@vanij.com"},
]
description = "A MCP server that provides an API to interact with WhatsApp"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0",
    "Pillow>=9.0.0",
    "click>=8.0.0",
    "mcp>=1.0.0",
    "whatsapp-api-client-python==0.0.49",
    "python-dotenv>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/vanij-platform/whatsapp-mcp-server"

[project.scripts]
whatsapp-mcp = "whatsapp_mcp.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.25.3",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "isort>=5.0.0",
    "black>=23.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.pytest]
testpaths = ["tests"]
